Did some testing
Former-commit-id: 8a0b60ddab72d177bdd8ced5227f4da7563889b8

+++ project/GridWorld/episodeplayergw.h
@@ -18,7 +18,7 @@ public:
     EpisodePlayerGW(string mapTag,vector<vector<float>> sequence);
     void initMap();
     void showMap();
-    void showScores(vector<vector<int>> scores);
+    void showScores(vector<vector<float>> scores);
     void playEpisode();
     MapGW getMap();
     vector<vector<float>> getSequence();
